+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - LiveViewNative.SwiftUI.Client
 - `NavigationLink` can perform a replace navigation by adding the `phx-replace` attribute
+- `NavigationLink` takes a `destination` template to customize the connecting phase View for its navigation event.
 
 ### Changed